refactor: Extract rename dialog
