+++ app/login/components/profile-tables.tsx
@@ -1,13 +1,7 @@
 import { t, Trans } from "@lingui/macro";
-import { LoadingButton, TabContext } from "@mui/lab";
 import {
   Box,
   Button,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogProps,
-  DialogTitle,
   Link,
   Skeleton,
   styled,
@@ -19,28 +13,25 @@ import {
   tableHeadClasses,
   TableRow,
   tableRowClasses,
-  TextField,
   Typography,
-  useEventCallback,
 } from "@mui/material";
 import { PUBLISHED_STATE } from "@prisma/client";
 import sortBy from "lodash/sortBy";
 import NextLink from "next/link";
-import React, { FormEvent, useState } from "react";
+import React from "react";
 
-import { VisualizeTab, VisualizeTabList } from "@/components/tabs";
 import useDisclosure from "@/components/use-disclosure";
 import { ParsedConfig } from "@/db/config";
 import { sourceToLabel } from "@/domain/datasource";
 import { truthy } from "@/domain/types";
 import { useUserConfigs } from "@/domain/user-configs";
 import { useDataCubesMetadataQuery } from "@/graphql/hooks";
-import { Locale } from "@/locales/locales";
 import { useRootStyles } from "@/login/utils";
 import { useLocale } from "@/src";
 import { removeConfig, updateConfig } from "@/utils/chart-config/api";
 import { useMutate } from "@/utils/use-fetch-data";
 
+import { RenameDialog } from "../../components/rename-dialog";
 import { ActionProps, RowActions } from "../../components/row-actions";
 
 const PREVIEW_LIMIT = 3;
@@ -173,145 +164,6 @@ export const ProfileVisualizationsTable = (
 type ProfileVisualizationsRowProps = {
   userId: number;
   config: ParsedConfig;
-};
-
-const RenameDialog = ({
-  config,
-  locale,
-  onClose,
-  userId,
-  ...props
-}: {
-  config: ParsedConfig;
-  locale: Locale;
-  onClose: () => void;
-  userId: number;
-} & Omit<DialogProps, "onClose">) => {
-  const { invalidate: invalidateUserConfigs } = useUserConfigs();
-  const [renameIndex, setRenameIndex] = useState(0);
-
-  const updateConfigMut = useMutate(updateConfig);
-
-  const handleRename = useEventCallback(
-    async (ev: FormEvent<HTMLFormElement>) => {
-      const arrData = Array.from(new FormData(ev.currentTarget));
-      const formData = arrData.reduce((acc, [key, value]) => {
-        const [_field, indexS, lang] = key.split(".");
-        const index = Number(indexS);
-        acc[index] = acc[index] || ({} as Record<string, string>);
-        acc[index][lang as Locale] = `${value}`;
-        return acc;
-      }, [] as Record<Locale, string>[]);
-      ev.preventDefault();
-
-      await updateConfigMut.mutate({
-        key: config.key,
-        data: {
-          ...config.data,
-          chartConfigs: config.data.chartConfigs.map((x, i) => ({
-            ...x,
-            meta: {
-              ...x.meta,
-              title: formData[i],
-            },
-          })),
-        },
-      });
-
-      invalidateUserConfigs();
-      onClose?.();
-    }
-  );
-
-  return (
-    <Dialog {...props} onClose={onClose} fullWidth>
-      <form style={{ display: "contents" }} onSubmit={handleRename}>
-        <DialogTitle>
-          <Trans id="profile.chart.rename-dialog.title">
-            Rename the visualization
-          </Trans>
-        </DialogTitle>
-        <DialogContent>
-          <Typography variant="body2">
-            <Trans id="profile.chart.rename-dialog.description">
-              Enhance chart clarity with a clear title; a good title helps
-              understanding chart content.
-            </Trans>
-          </Typography>
-          <TabContext value={`${renameIndex}`}>
-            <VisualizeTabList
-              sx={{ my: "1rem" }}
-              onChange={(_ev, newTab) => setRenameIndex(newTab)}
-            >
-              {config.data.chartConfigs.map((x, i) => {
-                return (
-                  <VisualizeTab
-                    key={i}
-                    value={`${i}`}
-                    label={
-                      <span>
-                        {x.meta.title[locale] !== ""
-                          ? x.meta.title[locale]
-                          : t({ id: "annotation.add.title" })}
-                      </span>
-                    }
-                  />
-                );
-              })}
-            </VisualizeTabList>
-            {config.data.chartConfigs.map((x, i) => {
-              return (
-                <Box
-                  key={i}
-                  sx={{
-                    padding: 0,
-                    gap: "1rem",
-                    flexDirection: "column",
-                    display: `${renameIndex}` === `${i}` ? "flex" : "none",
-                  }}
-                >
-                  <TextField
-                    name={`title.${i}.de`}
-                    label={t({ id: "controls.language.german" })}
-                    defaultValue={x.meta.title.de}
-                  />
-                  <TextField
-                    name={`title.${i}.fr`}
-                    label={t({ id: "controls.language.french" })}
-                    defaultValue={x.meta.title.fr}
-                  />
-                  <TextField
-                    name={`title.${i}.it`}
-                    label={t({ id: "controls.language.italian" })}
-                    defaultValue={x.meta.title.it}
-                  />
-                  <TextField
-                    name={`title.${i}.en`}
-                    label={t({ id: "controls.language.english" })}
-                    defaultValue={x.meta.title.en}
-                  />
-                </Box>
-              );
-            })}
-          </TabContext>
-        </DialogContent>
-        <DialogActions sx={{ pb: 6, pr: 6 }}>
-          <Button variant="outlined" onClick={onClose}>
-            {t({ id: "login.rename-dialog.cancel" })}
-          </Button>
-          <LoadingButton
-            sx={{ minWidth: "auto" }}
-            loading={updateConfigMut.status === "fetching"}
-            variant="contained"
-            color="primary"
-            type="submit"
-          >
-            {t({ id: "login.rename-dialog.save" })}
-          </LoadingButton>
-        </DialogActions>
-      </form>
-    </Dialog>
-  );
 };
 
 const ProfileVisualizationsRow = (props: ProfileVisualizationsRowProps) => {
